Add Auth (JWT) and Improve Models
Added authentication using `simpleJWT`. Note a lot of cruft remains in
settings.py related to the CORS headers for this.

Also improved Transaction model by adding a `tax_deduction` boolean
value (also added in serializers etc.).

Added a `Balance` model to record the current balance of an
Account. Also added serializers and view (read-only), and urls for this.

Improved handling of `Transaction.save()` method by moving logic to update
balances into a separate function (`_update_balances()`), and by
amending this logic to handle cases where the Account, as well as the
amount, is changed.

Made revisions of a lot of docstrings to improve compliance with
PEP257 (etc.)

Signed-off-by: Rod Manning <rod.t.manning@gmail.com>